refactor(llm_agent): early return for None pos in _build_observation, move test
Refactor _build_observation() to handle self.pos is None with a single
early return at the top of the vision block, instead of duplicating
the guard in each grid/space branch.

Move test_generate_obs_with_none_pos from standalone
test_agent_observation_none_pos.py into test_llm_agent.py.
Delete the standalone file per review feedback.
